Unify Pinboard and FauxPinboard into single class
* Replace old system of separate Pinboard and FauxPinboard classes with
  a single Pinboard class that can operate in Mock or Production mode
* Rewrite node-pinboard basically entirely
* Use imo nicer API for Pinboard, e.g. Pinboard.tags.get rather than
  Pinboard.getTags
* Remove node-pinboard tests that aren't working now anyway, and I
  don't want to go figure out tests, yes I do feel bad about this
* Simplify usePinboard hook
* Fix type errors that I couldn't fix before the two classes were
  unified

One further note:
I removed node-pinboard's RequestParams and related typing. I think
this information is now captured in method parameters, but it's
possible there was something I don't yet understand about TypeScript
that this was doing which I have now removed. If that's the case, sorry
to future me who will surely be mad reading this.